backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
cloud-sql-python-connector==1.9.0
pg8000==1.30.5
google-cloud-secret-manager==2.17.0
google-cloud-aiplatform==1.38.1
firebase-admin==6.3.0
stripe==7.8.0
pillow==10.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pydantic-settings==2.1.0
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
pytest-cov==4.1.0
factory-boy==3.3.0
